# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BackSpyne operator and local-node telemetry API
 * OpenAPI spec version: 0.2.0
 */

export interface HealthStatus {
  status: string;
}