# lib/api-zod/src/generated/types/scanSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BackSpyne operator and local-node telemetry API
 * OpenAPI spec version: 0.2.0
 */
import type { ScanSessionStatus } from './scanSessionStatus';

export interface ScanSession {
  id: string;
  ownerId: string;
  label: string;
  status: ScanSessionStatus;
  startedAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  /** @nullable */
  consentNote?: string | null;
}